feat(cli): add --list-editions argument and update README
This commit introduces a new command-line argument, --list-editions, which allows users to list available Windows editions in an ISO or WIM file. The new functionality requires the user to provide a file path as input, and then the list_editions() function processes the file to display the available Windows editions.  The README file has been updated to include instructions on how to use the --list-editions argument, along with an example command..